Fix fatal error in metabox saving

+++ includes/metabox.php
@@ -169,12 +169,11 @@ function edd_cr_save_meta_data( $post_id ) {
         // Grab the items this post was previously restricted to and remove related meta
         $previous_items = get_post_meta( $post_id, '_edd_cr_restricted_to', true );
 
-        foreach( $previous_items as $item ) {
-
-             if( 'any' !== $item['download'] ) {
-
-                delete_post_meta( $item['download'], '_edd_cr_protected_post', $post_id );
-
+        if( $previous_items );
+            foreach( $previous_items as $item ) {
+                 if( 'any' !== $item['download'] ) {
+                    delete_post_meta( $item['download'], '_edd_cr_protected_post', $post_id );
+                 }
             }
         }
